// ecis-nodejs-yseq-feat_soap1218/service/ecissdk/model/data.ts
export interface MysqlSelResp {
  result: string
  data: { rows: any[] }
  msg?: string
}